Improve packet logging.  It now uses a typical hexdump style content dump.  Ensure packets are logged at every datalink receive and transmit.  (TODO: DDCMP UDP mode, SIMHdmc.)
Fix daemon mode which, at least on Linux, doesn't work if there are
threads, so the timer wheel has to be started afterwards.

Fix a corner case in sending routing updates where the holdoff flag
was not being cleared and consequently subsequent updates would not be
scheduled.  This should fix the problem seen at PYRTR.


git-svn-id: svn://akdesign.dyndns.org/pydecnet/trunk@461 02bf3f84-b523-46f3-a765-a93b7296b289